fix: check if column exists

+++ packages/data-store/src/migrations/6.addAttachments.ts
@@ -10,13 +10,18 @@ export class AddAttachments1753543810823 implements MigrationInterface {
     public async up(queryRunner: QueryRunner): Promise<void> {
         debug(`Running migration 'AddAttachments1753543810823'`)
 
-        debug(`adding 'attachments' column to 'message' table`)
-        await queryRunner.addColumn(migrationGetExistingTableByName(queryRunner, 'message'), new TableColumn({
-            name: "attachments",
-            type: "text",
-            isNullable: true,
-        }))
-        debug(`added 'attachments' column to 'message' table`)
+        const table = migrationGetExistingTableByName(queryRunner, 'message');
+        const hasColumn = table?.findColumnByName('attachments');
+
+        if (!hasColumn) {
+            debug(`adding 'attachments' column to 'message' table`)
+            await queryRunner.addColumn(table, new TableColumn({
+                name: "attachments",
+                type: "text",
+                isNullable: true,
+            }))
+            debug(`added 'attachments' column to 'message' table`)
+        }
 
         debug(`Migration 'AddAttachments1753543810823' completed successfully`)
     }
@@ -24,9 +29,14 @@ export class AddAttachments1753543810823 implements MigrationInterface {
     public async down(queryRunner: QueryRunner): Promise<void> {
         debug(`Rolling back migration 'AddAttachments1753543810823'`)
 
-        debug(`removing 'attachments' column from 'message' table`)
-        await queryRunner.dropColumn(migrationGetExistingTableByName(queryRunner, 'message'), 'attachments')
-        debug(`removed 'attachments' column from 'message' table`)
+        const table = migrationGetExistingTableByName(queryRunner, 'message');
+        const hasColumn = table?.findColumnByName('attachments');
+
+        if (hasColumn) {
+            debug(`removing 'attachments' column from 'message' table`)
+            await queryRunner.dropColumn(migrationGetExistingTableByName(queryRunner, 'message'), 'attachments')
+            debug(`removed 'attachments' column from 'message' table`)
+        }
 
         debug(`Migration 'AddAttachments1753543810823' rolled back successfully`)
     }
